requirements.txt
python-telegram-bot==20.7
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0